'''Compilation rosetta'''
 * Compilation works but linkage fails ...

+++ trunk/projects/Proteins/Rosetta/ProteinMover/RigidBodyTransMover.h
@@ -16,7 +16,7 @@
 #  include <core/kinematics/MoveMap.hh>
 #  include <core/kinematics/FoldTree.hh>
 #  include <core/conformation/Conformation.hh>
-#  include <protocols/init.hh>
+//#  include <protocols/init.hh>
 #  include <protocols/moves/BackboneMover.hh>
 #  include <protocols/moves/Mover.hh>
 #  include <protocols/moves/MoverCreator.hh>
@@ -93,7 +93,7 @@ public:
 
       // Set a jump
       core::kinematics::FoldTree foldTree = pose->fold_tree();
-      int cutpointResidue = std::floor(((double)firstResidue + (double)secondResidue) / 2.0);
+      int cutpointResidue = (int)std::floor(((double)firstResidue + (double)secondResidue) / 2.0);
 
       foldTree.new_jump(firstResidue, secondResidue, cutpointResidue);
       pose->fold_tree(foldTree);
